refactor(components): refactor upload
