+++ packages/components/upload/__tests__/upload-dragger.spec.ts
@@ -1,5 +1,6 @@
 import { provide, h, defineComponent } from 'vue'
 import makeMount from '@element-plus/test-utils/make-mount'
+import { uploadContextKey } from '@element-plus/tokens'
 import UploadDragger from '../src/upload-dragger.vue'
 
 const AXIOM = 'Rem is the best girl'
@@ -9,7 +10,7 @@ const Wrapper = defineComponent({
     onDrop: Function,
   },
   setup(props, { slots }) {
-    provide('uploader', { accept: 'video/*' })
+    provide(uploadContextKey, { accept: 'video/*' })
     return () => h(UploadDragger, props, slots)
   },
 })
