+++ packages/components/upload/index.ts
@@ -1,13 +1,10 @@
-import Upload from './src/index.vue'
+import { withInstall } from '@element-plus/utils/with-install'
+import Upload from './src/upload.vue'
 
-import type { App } from 'vue'
-import type { SFCWithInstall } from '@element-plus/utils/types'
+export const ElUpload = withInstall(Upload)
+export default ElUpload
 
-Upload.install = (app: App): void => {
-  app.component(Upload.name, Upload)
-}
-
-const _Upload = Upload as SFCWithInstall<typeof Upload>
-
-export default _Upload
-export const ElUpload = _Upload
+export * from './src/upload'
+export * from './src/upload-content'
+export * from './src/upload-list'
+export * from './src/ajax'
