+++ packages/tokens/index.ts
@@ -6,3 +6,4 @@ export * from './config-provider'
 export * from './radio'
 export * from './tabs'
 export * from './scrollbar'
+export * from './upload'
