+++ packages/tokens/upload.ts
@@ -1,0 +1,8 @@
+import type { InjectionKey } from 'vue'
+
+export interface UploadContext {
+  accept: string
+}
+
+export const uploadContextKey: InjectionKey<UploadContext> =
+  Symbol('uploadContextKey')
